crawl/requirements.txt
beautifulsoup4==4.11.1
ddddocr==1.4.7
demjson==1.6
httpx==0.23.0
loguru==0.6.0
lxml==4.9.1
pandas==1.5.1
psycopg2==2.9.5
pycryptodome==3.15.0
pymongo==4.3.2
python_dateutil==2.8.2
requests==2.28.1
selenium==4.5.0
